+++ app/api/auth/[...nextauth]/route.tsx
@@ -1,22 +1,5 @@
-
 import NextAuth from 'next-auth'
-import GoogleProvider from 'next-auth/providers/google'
-import type { Adapter } from 'next-auth/adapters';
-import { PrismaAdapter } from "@auth/prisma-adapter"
-import { PrismaClient } from "@prisma/client"
-
-const prisma = new PrismaClient()
-
-export const authOptions = {
-  secret: process.env.NEXT_PUBLIC_NEXTAUTH_SECRET,
-  adapter: PrismaAdapter(prisma) as Adapter,
-  providers: [
-    GoogleProvider({
-      clientId: process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID ?? '',
-      clientSecret: process.env.NEXT_PUBLIC_GOOGLE_CLIENT_SECRET ?? '',
-    }),
-  ],
-}
+import { authOptions } from '@/app/utils/authOptions';
 
 const handler = NextAuth(authOptions)
 console.log(handler);
