+++ app/utils/authOptions.ts
@@ -1,0 +1,17 @@
+import { PrismaAdapter } from "@auth/prisma-adapter";
+import { PrismaClient } from "@prisma/client";
+import { Adapter } from "next-auth/adapters";
+import GoogleProvider from 'next-auth/providers/google'
+
+const prisma = new PrismaClient()
+
+export const authOptions = {
+  secret: process.env.NEXT_PUBLIC_NEXTAUTH_SECRET,
+  adapter: PrismaAdapter(prisma) as Adapter,
+  providers: [
+    GoogleProvider({
+      clientId: process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID ?? '',
+      clientSecret: process.env.NEXT_PUBLIC_GOOGLE_CLIENT_SECRET ?? '',
+    }),
+  ],
+}
